+++ package.json
@@ -26,6 +26,7 @@
     "next-head-seo": "^0.1.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "recoil": "^0.7.7",
     "swr": "^2.1.0",
     "typescript": "4.9.5"
   },
