+++ src/styles/globals.css
@@ -3,13 +3,3 @@
   padding: 0;
   margin: 0;
 }
-html,
-body {
-  max-width: 100vw;
-  overflow-x: hidden;
-}
-body {
-  color: rgb(var(--foreground-rgb));
-  background: linear-gradient(to bottom, transparent, rgb(var(--background-end-rgb)))
-    rgb(var(--background-start-rgb));
-}
